api/requirements.txt

Django==5.2.7
djangorestframework==3.16.1
psycopg==3.2.11
gunicorn==23.0.0
python-dotenv==1.1.1
django-cors-headers==4.9.0
Unidecode==1.4.0
